+++ internal_filesystem/lib/drivers/display/usb_disp/__init__.py
@@ -1,0 +1,17 @@
+from . import usb_disp
+
+__all__ = [
+    'USB_DISP',
+    'STATE_HIGH',
+    'STATE_LOW',
+    'STATE_PWM',
+    'BYTE_ORDER_RGB',
+    'BYTE_ORDER_BGR',
+]
+
+USB_DISP = usb_disp.USB_DISP
+STATE_HIGH = usb_disp.STATE_HIGH
+STATE_LOW = usb_disp.STATE_LOW
+STATE_PWM = usb_disp.STATE_PWM
+BYTE_ORDER_RGB = usb_disp.BYTE_ORDER_RGB
+BYTE_ORDER_BGR = usb_disp.BYTE_ORDER_BGR